Move getRisk to ComplianceTable dir
